+++ gwe/di.py
@@ -33,6 +33,7 @@ _LOG = logging.getLogger(__name__)
 SpeedStepChangedSubject = NewType('SpeedStepChangedSubject', Subject)
 FanProfileChangedSubject = NewType('FanProfileChangedSubject', Subject)
 OverclockProfileChangedSubject = NewType('OverclockProfileChangedSubject', Subject)
+SettingChangedSubject = NewType('SettingChangedSubject', Subject)
 MainBuilder = NewType('MainBuilder', Gtk.Builder)
 EditFanProfileBuilder = NewType('EditFanProfileBuilder', Gtk.Builder)
 EditOverclockProfileBuilder = NewType('EditOverclockProfileBuilder', Gtk.Builder)
@@ -117,6 +118,11 @@ class ProviderModule(Module):
     @provider
     def provide_overclock_profile_changed_subject(self) -> OverclockProfileChangedSubject:
         return OverclockProfileChangedSubject(Subject())
+
+    @singleton
+    @provider
+    def provide_overclock_profile_changed_subject(self) -> SettingChangedSubject:
+        return SettingChangedSubject(Subject())
 
 
 INJECTOR = Injector(ProviderModule)